+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# v0.2.0 (2019-01-11)
 - Refactor `default_backend` to return a `Box<dyn Backend>`.
 - Check the `DIALOG` and `DISPLAY` environment variables in `default_backend`.
 - Add the `Stdio` backend.